+++ source/common/vue/PopoverWrapper.vue
@@ -8,11 +8,42 @@
   </Teleport>
 </template>
 <script setup lang="ts">
-import { onBeforeUnmount, onMounted, onUpdated, ref } from 'vue'
+import { onBeforeUnmount, onMounted, onUpdated, ref, computed } from 'vue'
 
-const props = defineProps<{ target: HTMLElement }>()
+type Placement = 'above'|'below'|'left'|'right'
+
+const props = defineProps<{
+  target: HTMLElement
+  placementPriorities?: Placement[]
+}>()
 
 const emit = defineEmits<(e: 'close') => void>()
+
+const placementPriorities = computed<[Placement, Placement, Placement, Placement]>(() => {
+  if (props.placementPriorities === undefined) {
+    return [ 'below', 'right', 'above', 'left' ] as [Placement, Placement, Placement, Placement]
+  }
+
+  let placements = [...new Set(props.placementPriorities)]
+
+  // Ensure it includes all elements once
+  if (!placements.includes('below')) {
+    placements.push('below')
+  }
+  if (!placements.includes('right')) {
+    placements.push('right')
+  }
+
+  if (!placements.includes('above')) {
+    placements.push('above')
+  }
+
+  if (!placements.includes('left')) {
+    placements.push('left')
+  }
+
+  return placements as [Placement, Placement, Placement, Placement]
+})
 
 const popupWrapper = ref<HTMLDivElement|null>(null)
 const popupArrow = ref<HTMLDivElement|null>(null)
@@ -102,9 +133,11 @@ function place (): void {
 
   const elemRect = props.target.getBoundingClientRect()
 
+  // Where we should align the Popover to
   let x = elemRect.left + elemRect.width / 2
   let y = elemRect.top + elemRect.height
 
+  // Popover width and height
   let height = el.offsetHeight
   let width = el.offsetWidth
 
@@ -114,8 +147,32 @@ function place (): void {
   const right = window.innerWidth - left - elemRect.width
   const bottom = window.innerHeight - top - elemRect.height
 
+  let canPlaceBelow = bottom > height + 10 || elemRect.top < 50
+  let canPlaceRight = right > width + 10 && height <= window.innerHeight - 20 - y
+  let canPlaceAbove = top > height + 10
+  let canPlaceLeft = left + 10 < width && height <= window.innerHeight - 20 - y
+
+  let actualPlacement: Placement = 'below'
+
+  // First priority that can be fulfilled wins
+  for (const prio of placementPriorities.value) {
+    if (prio === 'above' && canPlaceAbove) {
+      actualPlacement = 'above'
+      break
+    } else if (prio === 'below' && canPlaceBelow) {
+      actualPlacement = 'below'
+      break
+    } else if (prio === 'left' && canPlaceLeft) {
+      actualPlacement = 'left'
+      break
+    } else if (prio === 'right' && canPlaceRight) {
+      actualPlacement = 'right'
+      break
+    }
+  }
+
   // 10px: arrow plus the safety-margin
-  if (bottom > height + 10 || elemRect.top < 50) {
+  if (actualPlacement === 'below') {
     // Below element
     arrow.classList.add('up')
     el.style.top = `${y + arrowSize}px` // 5px margin for arrow
@@ -146,7 +203,7 @@ function place (): void {
       el.style.height = `${window.innerHeight - 20 - y}px`
       height = el.offsetHeight
     }
-  } else if (right > width + 10 && height <= window.innerHeight - 20 - y) {
+  } else if (actualPlacement === 'right') {
     // We can place it right of the element
     // Therefore re-compute x and y
     x = elemRect.left + elemRect.width
@@ -170,7 +227,7 @@ function place (): void {
     } if (el.offsetTop < 10) {
       el.style.top = '10px'
     }
-  } else {
+  } else if (actualPlacement === 'above') {
     // Above
     // Therefore re-compute x and y
     x = elemRect.left + elemRect.width / 2
@@ -194,6 +251,29 @@ function place (): void {
     } if (el.offsetLeft < 10) {
       el.style.left = '10px'
     }
+  } else if (actualPlacement === 'left') {
+    // We can place it left of the element
+    x = elemRect.left - width
+    y = elemRect.top + elemRect.height / 2
+    el.style.left = `${x - arrowSize}px`
+    if (y + height / 2 > window.innerHeight - arrowSize) {
+      el.style.top = `${window.innerHeight - height - arrowSize}px`
+    } else {
+      el.style.top = `${y - height / 2}px`
+    }
+
+    if (showArrow) {
+      arrow.classList.add('right')
+      arrow.style.left = `${left + width}px`
+      arrow.style.top = `${top + elemRect.height / 2 - arrow.offsetHeight / 2}px`
+    }
+
+    // Ensure the popup is completely visible (move inside the document if it's at an edge)
+    if (el.offsetTop + el.offsetHeight > window.innerHeight - 10) {
+      el.style.top = `${window.innerHeight - el.offsetHeight - 10}px`
+    } if (el.offsetTop < 10) {
+      el.style.top = '10px'
+    }
   }
 }
 </script>
@@ -210,35 +290,14 @@ function place (): void {
 @arrow-size: 20px;
 @color-light: rgb(230, 230, 230);
 @color-dark: rgb(30, 30, 30);
+@border-color-light: rgb(213, 213, 213);
+@border-color-dark: rgb(70, 70, 70);
 
 /* * * * * * * * * * * * * *
  *                         *
  *    END VARIABLES AREA   *
  *                         *
  * * * * * * * * * * * * * */
-
-/* * * * * * * * * * * * * *
- *                         *
- *     BEGIN MIXIN AREA    *
- *                         *
- * * * * * * * * * * * * * */
-
-// Makes a div flex, can be used to generate a grid style button set
-.flex-button-mixin(@size) {
-  flex-grow: 1;
-  width: @size;
-  height: @size;
-  line-height: @size;
-  text-align: center;
-  transition: 0.2s all ease;
-  cursor: default;
-}
-
- /* * * * * * * * * * * * * *
-  *                         *
-  *      END MIXIN AREA     *
-  *                         *
-  * * * * * * * * * * * * * */
 
 body .popover-arrow {
   width: 0;
@@ -249,28 +308,28 @@ body .popover-arrow {
   &.up {
     border-left: @arrow-size solid transparent;
     border-right: @arrow-size solid transparent;
-    border-bottom: @arrow-size solid @color-light;
+    border-bottom: @arrow-size solid @border-color-light;
   }
   &.down {
     border-left: @arrow-size solid transparent;
     border-right: @arrow-size solid transparent;
-    border-top: @arrow-size solid @color-light;
+    border-top: @arrow-size solid @border-color-light;
   }
   &.left {
     border-top: @arrow-size solid transparent;
     border-bottom: @arrow-size solid transparent;
-    border-right: @arrow-size solid @color-light;
+    border-right: @arrow-size solid @border-color-light;
   }
   &.right {
     border-top: @arrow-size solid transparent;
     border-bottom: @arrow-size solid transparent;
-    border-left: @arrow-size solid @color-light;
+    border-left: @arrow-size solid @border-color-light;
   }
 }
 
 body .popover {
   background-color: @color-light;
-  border: 1px solid rgb(213, 213, 213);
+  border: 1px solid @border-color-light;
   box-shadow: 0px 0px 40px 0px rgba(0, 0, 0, .4);
   border-radius: 8px;
   position: fixed;
@@ -347,159 +406,21 @@ body .popover {
     height: 10px;
     margin: 1px;
   }
-
-  .btn-share {
-    .flex-button-mixin(@size-share-buttons);
-    background-position: center center;
-    background-size: contain;
-    background-repeat: no-repeat;
-
-    &.htm {
-    //  background-image: url('../../img/html5.svg');
-      &:hover {
-        /* W3C color code */
-        background-color: rgb(78, 175, 87);
-      }
-    }
-
-    &.pdf {
-    //  background-image: url('../../img/pdf.svg');
-      &:hover {
-        /* Adobe reader PDF color code */
-        background-color: lighten(rgb(253, 48, 26), 10%);
-      }
-    }
-
-    &.docx {
-    //  background-image: url('../../img/docx.svg');
-      &:hover {
-        /* MS Word color code */
-        background-color: lighten(rgb(45, 88, 151), 10%);
-      }
-    }
-
-    &.odt {
-    //  background-image: url('../../img/odt.svg');
-    }
-
-    &.odt, &.tex, &.plain {
-      color: rgba(20, 20, 20, 1);
-      &:hover { background-color: rgba(220, 220, 220, 1); color:black; }
-    }
-
-    &.revealjs {
-      color: rgba(120, 120, 120, 1);
-      &:hover { background-color: rgba( 40,  40,  40, 1); color: white; }
-    }
-
-    // revealJS theme buttons
-    &.revealjs-beige {
-      color: #333333;
-      background-color: #f7f2d3;
-      &:hover { background-color: darken(#f7f2d3, 20%); color: #333333; }
-    }
-    &.revealjs-black {
-      color: #ffffff;
-      background-color: #222222;
-      &:hover { background-color: lighten(#222222, 20%); color: #ffffff; }
-    }
-    &.revealjs-league {
-      color: #71e9f4;
-      background-color: #555a5f;
-      &:hover { background-color: lighten(#555a5f, 20%); color: lighten(#71e9f4, 20%); }
-    }
-    &.revealjs-moon {
-      color: #268bd2;
-      background-color: #002b36;
-      &:hover { background-color: lighten(#002b36, 20%); color: lighten(#268bd2, 20%); }
-    }
-    &.revealjs-serif {
-      color: #383d3d;
-      background-color: #f0f1eb;
-      &:hover { background-color: darken(#f0f1eb, 10%); color: #383d3d; }
-    }
-    &.revealjs-sky {
-      color: #333333;
-      background-color: #add9e4;
-      &:hover { background-color: darken(#add9e4, 10%); color: #333333; }
-    }
-    &.revealjs-solarized {
-      color: #657b83;
-      background-color: #fdf6e3;
-      &:hover { background-color: darken(#fdf6e3, 10%); color: #657b83; }
-    }
-    &.revealjs-white {
-      color: #222222;
-      background-color: white;
-      &:hover { background-color: darken(white, 20%); color: #222222; }
-    }
-    // END revealJS theme buttons
-
-    &.rst {
-      color: rgb(28, 78, 98);
-      &:hover { color: white; background-color: rgb(28, 78, 98); } // This is literally the first colour I found on the web for rst
-    }
-
-    &.rtf {
-      color: rgb(223, 51, 68);
-      &:hover { background-color: rgb(223, 51, 68); color: white; }
-    }
-
-    &.org {
-      color: lighten(rgb(0, 42, 53), 10%);
-      &:hover { background-color: rgb(0, 42, 53); color: white; }
-    }
-
-    &.textbundle {
-      color: #00b58c;
-      &:hover { background-color: #00b58c; color:white; }
-    }
-
-    &.textpack {
-      color: lighten(#0b0080, 20%);
-      &:hover { background-color: #0b0080; color: white; }
-    }
-  }
-
-  // For the icon selector
-  .icon-block {
-    .flex-button-mixin(@size-icon-buttons);
-    font-size: 10px; // The clarity icons are way too big
-  }
-
-  // The revealjs theme selector's visibility
-  #reveal-themes {
-    transition: 0.2s height ease;
-    height: @size-share-buttons * 2;
-    overflow: hidden;
-    &.hidden { height: 0px; }
-  }
-
-  // Footnote edit popup
-  .footnote-edit {
-    textarea {
-      min-height: 150px;
-      border: none;
-      width: 250px; // Maximum width
-
-      &:focus { outline: 0; }
-    }
-  }
 }
 
 body.dark .popover-arrow {
-  &.up { border-bottom-color: @color-dark; }
-  &.down { border-top-color: @color-dark; }
-  &.left { border-right-color: @color-dark; }
-  &.right { border-left-color: @color-dark; }
+  &.up { border-bottom-color: @border-color-dark; }
+  &.down { border-top-color: @border-color-dark; }
+  &.left { border-right-color: @border-color-dark; }
+  &.right { border-left-color: @border-color-dark; }
 }
 
 body.dark .popover {
   background-color: @color-dark;
-  border: 1px solid rgb(70, 70, 70);
+  border: 1px solid @border-color-dark;
 
   hr {
-    border-color: rgb(70, 70, 70);
+    border-color: @border-color-dark;
   }
 }